1.0.0

+++ lib/renderers/webgl/utils/webgl_sprite_batch.dart
@@ -235,7 +235,7 @@ class WebGLSpriteBatch {
 
     // get the tilingSprites current alpha
     var alpha = tilingSprite.worldAlpha;
-    var tint = tilingSprite.tint;
+    var tint = tilingSprite.tint.toDouble();
 
     var verticies = this.vertices;
